requirements.txt
aiohttp==1.0.5
appdirs==1.4.3
async-timeout==1.2.0
beautifulsoup4==4.5.3
cffi==1.10.0
chardet==2.3.0
discord.py==0.16.7
multidict==2.1.4
packaging==16.8
pycparser==2.17
PyNaCl==1.0.1
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
websockets==3.2
